[10.0][REF] pos_order_history_reprint: refactoring the code (it-projects-llc/addons-dev#454)
* [REF] refactoring pos_order_history_reprint

* [DOC] update doc,manifest

* [DOC] update index.rst for pos_orders_history module

* [CI] add pos_orders_hitory tests

* [CI] fix pos_orders_history test

* [FIX] qweb link in manifest

* [CI] add test for pos_orders_history_reprint

* [CI] add tests

* [FIX] css style

* [CI] fix test for pos_orders_history_return
