+++ pos_orders_history/views/pos_orders_history_template.xml
@@ -11,4 +11,10 @@
             <link rel="stylesheet" href="/pos_orders_history/static/src/css/pos.css"/>
         </xpath>
     </template>
+
+    <template id="demo_assets_backend" name="pos_orders_history demo assets" inherit_id="web.assets_backend">
+        <xpath expr="." position="inside">
+            <script type="text/javascript" src="/pos_orders_history/static/src/js/tour.js"></script>
+        </xpath>
+    </template>
 </odoo>
