+++ pos_orders_history_reprint/README.rst
@@ -1,0 +1,47 @@
+.. image:: https://img.shields.io/badge/license-LGPL--3-blue.png
+   :target: https://www.gnu.org/licenses/lgpl
+   :alt: License: LGPL-3
+
+============================
+ POS Orders History Reprint
+============================
+
+Reprint paid POS Orders with POS interface.
+
+Credits
+=======
+
+Contributors
+------------
+* Losev Artyom
+
+Sponsors
+--------
+* `IT-Projects LLC <https://it-projects.info>`__
+
+Maintainers
+-----------
+* `IT-Projects LLC <https://it-projects.info>`__
+
+      To get a guaranteed support
+      you are kindly requested to purchase the module
+      at `odoo apps store <https://apps.odoo.com/apps/modules/10.0/pos_orders_history_reprint/>`__.
+
+      Thank you for understanding!
+
+      `IT-Projects Team <https://www.it-projects.info/team>`__
+
+Further information
+===================
+
+Demo: http://runbot.it-projects.info/demo/pos-addons/10.0
+
+HTML Description: https://apps.odoo.com/apps/modules/10.0/pos_orders_history_reprint/
+
+Usage instructions: `<doc/index.rst>`_
+
+Changelog: `<doc/changelog.rst>`_
+
+Notifications on updates: `via Atom <https://github.com/it-projects-llc/pos-addons/commits/10.0/pos_orders_history_reprint.atom>`_, `by Email <https://blogtrottr.com/?subscribe=https://github.com/it-projects-llc/pos-addons/commits/10.0/pos_orders_history_reprint.atom>`_
+
+Tested on Odoo 10.0 e03a06b3e9c202d9a87bc4f9bfe38e2643b95b20
