+++ pos_orders_history_reprint/doc/changelog.rst
@@ -1,0 +1,9 @@
+.. _changelog:
+
+Updates
+=======
+
+`1.0.0`
+-------
+
+- init version
