+++ pos_orders_history_reprint/models/__init__.py
@@ -1,3 +1,3 @@
 # -*- coding: utf-8 -*-
 # License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl.html).
-from . import models
+from . import model
