+++ pos_orders_history_reprint/static/src/css/main.css
@@ -1,5 +1,6 @@
-.button.reprint {
-    left:150px;
-}
+/* Copyright 2018 Artem Losev
+   License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl.html).*/
 
-    
+.button.reprint {
+    left:115px;
+}
