+++ pos_orders_history_reprint/static/src/xml/main.xml
@@ -1,4 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
+<!-- Copyright 2018 Artem Losev
+     License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl.html). -->
 <templates id="template" xml:space="preserve">
 
     <t t-extend="OrdersHistoryScreenWidget">
