+++ pos_orders_history_reprint/views/template.xml
@@ -1,11 +1,19 @@
 <?xml version="1.0" encoding="utf-8"?>
+<!-- Copyright 2018 Dinar Gabbasov <https://it-projects.info/team/GabbasovDinar>
+     Copyright 2018 Artem Losev
+     License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl.html). -->
 <odoo>
-    <data>
-        <template id="assets_backend" name="pos_orders_history assets" inherit_id="point_of_sale.assets">
-            <xpath expr="." position="inside">
-                <script type="text/javascript" src="/pos_orders_history_reprint/static/src/js/main.js"></script>
-                <link rel="stylesheet" href="/pos_orders_history_reprint/static/src/css/main.css"/>
-            </xpath>
-        </template>
-    </data>
+    <template id="assets_backend" name="pos_orders_history assets" inherit_id="point_of_sale.assets">
+        <xpath expr="." position="inside">
+            <script type="text/javascript" src="/pos_orders_history_reprint/static/src/js/screens.js"></script>
+            <link rel="stylesheet" href="/pos_orders_history_reprint/static/src/css/main.css"/>
+        </xpath>
+    </template>
+
+    <template id="demo_assets_backend" name="pos_orders_history demo assets" inherit_id="web.assets_backend">
+        <xpath expr="." position="inside">
+            <script type="text/javascript" src="/pos_orders_history_reprint/static/src/js/tour.js"></script>
+        </xpath>
+    </template>
+
 </odoo>
